requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
astroid==3.0.2
async-timeout==4.0.3
attrs==23.1.0
black==23.12.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
dnacentersdk==2.6.10
fastjsonschema==2.19.0
Flask==3.0.0
fonttools==4.47.0
frozenlist==1.4.0
future==0.18.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mccabe==0.7.0
meraki==1.40.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
Pillow==10.1.0
platformdirs==4.1.0
prettyprint==0.1.5
pylint==3.0.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.28.0
requests-toolbelt==1.0.0
schedule==1.2.1
six==1.16.0
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
